backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.2
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
sentry-sdk==2.29.1
sqlparse==0.5.3
whitenoise==6.12.0
